+++ src/cli/mod.rs
@@ -2,6 +2,7 @@
 
 pub use crate::config::Config;
 
+pub mod exec;
 pub mod init;
 pub mod max;
 pub mod run;