+++ src/exec/mod.rs
@@ -1,0 +1,5 @@
+//! Execute script contracts.
+
+pub use script::Script;
+
+pub mod script;